Move useCamera into its own file
