+++ docs/components/guide-examples/display/coordinates/CartesianCoordinatesExample.tsx
@@ -2,8 +2,8 @@ import { Mafs, Coordinates } from "mafs"
 
 export default function Example() {
   return (
-    <Mafs height={300} zoom={true}>
-      <Coordinates.Cartesian subdivisions={10} />
+    <Mafs height={300}>
+      <Coordinates.Cartesian />
     </Mafs>
   )
 }
